pyproject.toml
[project]
name = "app-finanzas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "celery==5.4.0",
    "dj-database-url>=3.0.1",
    "django>=4.2.27",
    "gunicorn>=23.0.0",
    "psycopg[binary,pool]>=3.2.13",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "redis==5.0.8",
    "mail-parser>=4.1.4",
    "whitenoise>=6.11.0",
    "django-axes>=6.0.0",
    "requests>=2.32.5",
    "requests-oauthlib>=2.0.0",
    "splitwise>=1.3.0",
    "django-htmx>=1.26.0",
    "django-pwa>=2.0.1",
    "pydantic>=2.0.0",
    "pillow>=11.0.0",
    "llama-cloud-services>=0.6.88",
]
